Add initial setup for authentication service
This commit introduces the initial structure for an authentication service, including settings for a MySQL database and a Redis cache. It also includes a .gitignore file, Docker Compose file, and the new Maven Wrapper files to ensure builds are consistent across environments. A starter Spring Boot Application file and a related test setup are also created.🚀👷‍♀️