Add comments counter feature to Posts 💬
The Post model now includes a commentsCount field to keep track of the number of comments for each post. A PostCounterService 🛠️ was added for incrementing and decrementing this count. A separate PostCommentsService and accompanying implementation 📚 were also added to handle deleting all comments related to a specific post. Corresponding modifications were carried out in the tests too. Lastly, a few classes were relocated to more appropriate packages and respective refactoring was performed.